⚡️(project) build the frontend before the docker img
When using the bootstrap command, for more consistency, if frontend
dependencies changed, we need to build the frontend first to update the
yarn.lock file (that may be updated) and then build the Docker image
that enforces a frozen yarn.lock file. If we do not respect this
particular tasks ordering, the Docker image build will fail.