⚗(arnold) add tray package metadata files
⚠ EXPERIMENTAL ⚠

We need a way for arnold to pull a git repository and detect the tray
path to add an application to available apps. This is the purpose of the
arnold.yml file that should stand in the project's root repository.

The second file that should stand at the tray's root called tray.yml
should contain at least the application name and version.