⚡️(project) improve Makefile tasks logic
It appears more effective that every Django management command uses
'docker-compose exec' on a running stack instead of trying to start a
new stack using run. This requires that services are running prior to
executing the aforementioned management command.

Dockerize is now treated as a service that can be used as a single
tool for the whole stack; thus it's no longer required to install it in
the app service image.