✨(transfer) add application to import pages from a Google sheet
We need to automatically import all pages from the existing fun-mooc.fr
website. The data was exported from the old website and imported in a
Google Sheet where it was polished and completed manually. It must now
be imported in Richie so we use basic scripts to browse through the
Google sheet and create each page and plugin.