🍻(gimporter) convert and clean relative urls and media
The news articles are imported from fun-mooc.fr include a lot of links
that would be broken once we move to the new site. In this
commit, I replace urls targetting courses and organizations by their
new url. If a course is not present on the new site, I replace the url
by a link to the search page.