♻️(gimporter) inflate media files found in content
In the content of the old fun-mooc.fr site, there are some media that
we don't have on the new site. We need to extract these urls, download
the corresponding files from the old site, store them in Django filer,
and replace the link in the content by the new link.