♻️(project) add custom nginx image including static files
Static files were stored in AWS S3 but our new best practice is
to store them in a custom nginx image with the following advantages:
- fast deployments (no more collectstatic)
- better static files consistency
- good performance as well since files are served directly from the node

In order to test this new image, we refactored the docker-compose project
to run the production images in the same context as in production.

Dockerize is no more installed in the development image, and we use
the dockerize Docker image instead.