♻️(circleci) refactor workflows to work well with github status checks
We were only running workflows for projects that required it but in github,
status checks can only be marked as durably required or not. Workflows
had to be marked required when they were run and not required when they
were absent... This is not possible so pull requests would be blocked.

We choose to always run all workflows, mark them as required in github,
and run a dummy job if nothing has changed.