✨(circleci) allow overriding the Docker image name
The site name could not contain any "-" or "_" because it had to
simultaneously be the name of a python module and of a docker image.
In the case of the "demo" site, this name is not explicit enough
for the Docker image and the ideal name was "richie-demo".

We choose to allow overriding the Docker image name by using the backend
Python package name if present in setup.cfg.