♻️(circleci) check versions, tags and changelog also for pull requests
When making a release, if we fail to upgrade a version or the changelog
correctly, discovering it at the moment we tag the release is too late.
Everything should be checked on all pushes. For this we allow the second
argument (tag) to be null on the check_tag function and extract it from
the CHANGELOG file in this case to check that everything is in order.