♻️(aws) share the same S3 bucket to manage the state of all sites
The "terraform.tfstate" file was conflicting between sites. We solved
this by relocating this file to each site's folder thanks to the
"-backend-config=key" setting. Doing this, we found out that it is
possible to share the same bucket for the states of all sites, so
we took this opportunity as it simplifies the use of "bin/state" and
keeps everything in a "richie-site-factory" bucket in S3.