✨(Makefile) prevent obsolete translations
We want to be able to override some translations from dependencies,
namely richie, without having to override portions of Python or template
code when it is not necessary.

It is possible to add translations manually in the django.po file but
the makemessages command will mark them obsolete each time it is run
because the corresponding string is not found, neither in Python, nor in
the template code of the active site in richie-site-factory.

We want to keep all translations active. Reinstating obsolete translations
with a quick sed hack seems like a good deal.