✨(demo) force user.is_superuser on login when they have is_staff True
OpenEdx hawthorn doesn't send in the oauth response if the user is a
superuser, only if he is a staff member. For the demo we decided to set
a user as superuser if he is staff. For this we added a function in the
authentication pipeline executed once we are sure a user exists and
check if he is a staff member. If it's the case then we set the user as
a superuser and we save it.