🐛(Makefile) run i18n before starting the app container
The .mo files are gitignored. They are computed in the Dockerfile
as a step in the build. However, in development, docker-compose
mounts the project sources to /app and thus hides the .mo files
that were computed during the build. Running i18n before starting
the app container ensures that translation overrides are taken
into account for a new comer running "make bootstrap".