🔥(funcampus) disable cms page cache for non-staff users
While investigating on a bug that disconnects CMS admins when they
preview a modified page, we decided to change our user management
strategy. Most of our users will not have an account on richie. As a
result, the monkey patch that we applied to enable cache to non-staff
users is now useless.