✨(funcampus) add monkey patch to limit the browser cache TTL
By default, Django CMS send cached response with a
Cache-control:max-age value that reflects the server cache TTL.

The thing is : we can invalidate a server side cache entry, but we
cannot invalidate our client browser cache entries. That's why we want
to set a long TTL on the server side, but a much lower TTL on the
browser cache.

A new settings and a monkey patch has been introduced to handle this :
CMS_MAX_BROWSER_CACHE_TTL.