🐛(funcampus) fix demo following template overrides
Some template overrides are breaking the demo because some
placeholders are not available anymore. We need to override the
demo settings in order to be able to generate the demo site
successfully.

Note: the PAGES_INFO does not support merge update yet so we
override the whole "annex" definition. This can be simplified as
soon as the merge update is added on this setting in richie.