🐛(all) include version in CMS cache prefix to bust cache on deployments
When deploying a new version of the app, the cache was not reset so this
could lead to bugs in certain cases because the cached html and the new
js could be incompatible. Including the version in the CMS cache prefix
makes sure that each version of the app handles cache separately.