📈(funmooc) enable xiti tracking
Configure and initialize xiti tag in a generic manner.
First dispatch data of each page visited by an user. Furthermore, collect data
when an user clicks on an action button.
We also collect organizations mentionned on the current page and pass their code
to Xiti. In this way, we will be able to filter analytics data by organization.
To work, this code required smarttag.js > 5.21.0