🐛(funmooc) fix contributors filter by customizing only human name
When requested to change the filter to "Contributors" by the support
team, we renamed the machine names as well. It was working fine for
the filter functionality but the autocomplete was broken.

A reflexion to improve the engineering is underway in richie but in
the meantime we should avoid renaming machine names on filters.