✅(funmooc) fix template test
A test failed due to a wrong html parsing. As contents in our tests are in
french, we have to manage accentuated entities. To get a pretty string, when
using etree.tostring we need first to encode this string in "ISO8859-1"
then decode the result in "UTF-8"

This test failed just now, because the error ocurred when we check if
the course run's start date is contained in the response.content.
In french, august is the only month contaning accentuated characters (août)
so this test failed only in August