♻️(css) add missing modal styles import and variables
The upgrade to Richie 2.10.0 required adding an import for modal
styles, which were recently integrated in a separate file, to
the main stylesheets for all the richie sites.
We also needed to add color entries for "modal", which partly
replaces "search-filter-group-modal" and breaks the build for
sites which replace the default theme with their own.

This was not documented in upgrade.md and therefore forgotten
during the upgrade process.